relocating footer and menu types
