+++ src/features/dashboard/components/footer/DashBoardFooter.tsx
@@ -3,7 +3,7 @@ import { FaFacebook, FaTwitter, FaTiktok } from 'react-icons/fa';
 import { DashBoardFooterStyles } from '@/shared/utils/twind/styles';
 import { useTranslation } from '@/shared/services/redux/hooks/useTranslation';
 import { HealthCheck } from '@/features/healthcheck/components/HealthCheck';
-import { SocialLink, FooterLink } from '@/shared/types/footer.type';
+import { SocialLink, FooterLink } from '@/shared/types/dashboard/footer.type';
 import { HealthStatusDisplay } from '@/features/healthcheck/components/HealthStatusDisplay';
 
 const SOCIAL_LINKS: SocialLink[] = [
