+++ src/features/dashboard/components/header/DashBoardHeader.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ContainerHeader } from './components/ContainerHeader';
-import { ToggleButtonProps } from '@/shared/types/menu.type';
+import { ToggleButtonProps } from '@/shared/types/dashboard/menu.type';
 
 export const DashBoardHeader: React.FC<ToggleButtonProps> = ({
   setSidebarToggle,
