+++ src/features/dashboard/components/menu/DashBoardToggleMenuBar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ContainerMenuBar } from './components/ContainerMenuBar';
-import { MenuBarProps } from '@/shared/types/menu.type';
+import { MenuBarProps } from '@/shared/types/dashboard/menu.type';
 import { DashBoardMenuBarStyles } from '@/shared/utils/twind/styles';
 
 export const DashBoardToggleMenuBar: React.FC<MenuBarProps> = ({
