+++ src/shared/types/dashboard/footer.type.ts
